A lot of fixes

+++ WoodenECS/stdafx.h
@@ -6,7 +6,7 @@
 #pragma once
 
 #include "targetver.h"
-#include "stdint.h"
+#include <stdint.h>
 #define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
 
 
@@ -15,8 +15,7 @@
 #define WECS_USE using namespace wecs
 
 
-#define INVALID_HANDLE (size_t)(-1)
-
+#define INVALID_HANDLE (uint32_t)(-1)
 
 
 // reference additional headers your program requires here
